Skip args where args_ttype is None
The try-catch block at line 62 allows args_ttype to be None. Line 74 was updated to an f-string to convert None to str(None) to prevent the program from crashing.

Recently ran into this issue when testing a development environment which caused a None type to appear. Changing to an f-string mitigated this issue and allowed for enumeration to continue.